requirements.txt
attrs==25.4.0
blinker==1.9.0
cattrs==25.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
flatbuffers==25.9.23
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
jh2==5.0.10
Jinja2==3.1.6
MarkupSafe==3.0.3
niquests==3.16.1
numpy==2.4.1
openmeteo_requests==1.7.4
openmeteo_sdk==1.23.0
pandas==2.3.3
platformdirs==4.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
qh3==1.5.6
requests==2.32.5
requests-cache==1.2.1
retry-requests==2.0.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
url-normalize==2.2.1
urllib3==2.6.3
urllib3-future==2.15.901
waitress==3.0.2
wassima==2.0.3
Werkzeug==3.1.4
